+++ wutengine/src/plugins/mod.rs
@@ -1,3 +1,6 @@
+use winit::event::{DeviceEvent, DeviceId, WindowEvent};
+use wutengine_graphics::windowing::WindowIdentifier;
+
 use crate::command::Command;
 use crate::runtime::RuntimeInitializer;
 
@@ -7,4 +10,22 @@ pub trait WutEnginePlugin: 'static {
 
     /// Called once when the runtime has just been built, and is starting
     fn on_start(&mut self, _commands: &mut Command) {}
+
+    /// Called once for each raw window event returned by the windowing system (currently [winit])
+    fn on_window_event(
+        &mut self,
+        _window: &WindowIdentifier,
+        _event: &WindowEvent,
+        _commands: &mut Command,
+    ) {
+    }
+
+    /// Called once for each raw device event returned by the windowing system (currently [winit])
+    fn on_device_event(
+        &mut self,
+        _device: DeviceId,
+        _event: &DeviceEvent,
+        _commands: &mut Command,
+    ) {
+    }
 }