+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This file documents the most important changes for each released version.
 ## [v0.0.7]
 - Initial support for optional components in ECS queries
 - Camera component now influences rendering through its position (as determined by its Transform component) and its perspective settings
+- Engine plugins can now listen and respond to raw window and device events
 
 ## [v0.0.6]
 - Complete refactor of core ECS system. We now use archetypes. It was a lot of work.